refactor: mobilise the entire interface structure to the app directory
