+++ components.json
@@ -11,11 +11,11 @@
     "prefix": ""
   },
   "aliases": {
-    "components": "@/components",
-    "utils": "@/lib/utils",
-    "ui": "@/components/ui",
-    "lib": "@/lib",
-    "hooks": "@/hooks"
+    "components": "@/app/_ui",
+    "utils": "@/app/_ui/lib/utils",
+    "ui": "@/app/_ui/components",
+    "lib": "@/app/_ui/lib",
+    "hooks": "@/app/_ui/hooks"
   },
   "iconLibrary": "lucide"
 }
