+++ src/app/_ui/components/card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/app/_ui/lib/utils"
 
 const Card = React.forwardRef<
   HTMLDivElement,
