+++ src/app/_ui/components/sidebar.tsx
@@ -5,19 +5,19 @@ import { Slot } from "@radix-ui/react-slot"
 import { VariantProps, cva } from "class-variance-authority"
 import { PanelLeft } from "lucide-react"
 
-import { useIsMobile } from "@/hooks/use-mobile"
-import { cn } from "@/lib/utils"
-import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Separator } from "@/components/ui/separator"
-import { Sheet, SheetContent } from "@/components/ui/sheet"
-import { Skeleton } from "@/components/ui/skeleton"
+import { useIsMobile } from "@/app/_ui/hooks/use-mobile"
+import { cn } from "@/app/_ui/lib/utils"
+import { Button } from "@/app/_ui/components/button"
+import { Input } from "@/app/_ui/components/input"
+import { Separator } from "@/app/_ui/components/separator"
+import { Sheet, SheetContent } from "@/app/_ui/components/sheet"
+import { Skeleton } from "@/app/_ui/components/skeleton"
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "@/components/ui/tooltip"
+} from "@/app/_ui/components/tooltip"
 
 const SIDEBAR_COOKIE_NAME = "sidebar:state"
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7
