+++ src/app/_ui/components/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils"
+import { cn } from "@/app/_ui/lib/utils"
 
 function Skeleton({
   className,
